requirements.txt
feedparser
vaderSentiment
requests
eventregistry
python-dotenv
supabase
anthropic
